pkg: cmd: remove "version" command from pkg
The "version" command needs the non-exported `pkg/version`.
his makes the `commands` pkg unusable otuside this project,
that is violating the original design.

We need to untie the `commands` pkg from the unshipped `version` pkg;
it is still correct to NOT ship the latter just yet.

The version command is tied to this source tree, so it
makes sense to move it `cmd`, which we do in this patch.

No expected changes in behaviour.
